requirements.txt
django
djangorestframework
drf-yasg
pylint
pylint-django
pylint-quotes
